api_yamdb/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.0.12
distlib==0.3.6
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
filelock==3.10.0
gunicorn==20.0.4
identify==2.5.21
idna==3.4
iniconfig==2.0.0
nodeenv==1.7.0
packaging==23.0
platformdirs==3.1.1
pluggy==0.13.1
pre-commit==3.2.0
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.7.1
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.15
virtualenv==20.21.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
